Fix: playing of instrument without note.
When an instrument is played without a note, the current playing
sample should restart its envelope, keeping its position. So it's
important not to change the current instrument.

Also, the same behavior happens even if the instrument is invalid (and
no note is specified); on the other hand, if the instrument is invalid
and a note is specified, the current note is being cut.

Co-developed with @bryc, who also prepared the tests