Fix: fadeout volume speed was doubled
This is a hairy issue in the XM format. xm-form.txt explicitly mentions
a counter that starts from 65536, so there must be some truth to that.
This notwithstanding, all trackers agree that the correct speed should
be 32768. In particular:

ft2-clone:
https://github.com/8bitbubsy/ft2-clone/blob/79e4f7b5749f12866a3b63f460d3e51559f1e137/src/ft2_replayer.c#L347

MilkyTracker: double the value at load time
https://github.com/milkytracker/MilkyTracker/blob/e0652ecca58ff2d4fc89b9eaec5ec3cd015bea91/src/milkyplay/LoaderXM.cpp#L219-L221

libopenmpt:
https://github.com/OpenMPT/openmpt/blob/c827d43d55113cc1d2e60c5eb83a8e25f8aa43be/soundlib/XMTools.cpp#L61-L62

The attached test shows the problem, and the fixed version plays
exactly like OpenMPT and MilkyTracker, as expected.

Co-developed with @bryc, who also prepared the tests