Fix a couple of bugs in effect Rxy (multiretrig)
Rxy is possibly the most quirky command in the XM format, and it's
hard to get right in all its dark corners. This gets libxm closer
to FT2-Clone and XMPlay (even closer than MilkyTracker) but still
not quite perfect.

The main bug fixed here is that the volume change of a Rxy command only
takes effect if the volume column is empty, and only if there is no volume
envelope in the instrument (and in this case, the envelope is not even
restarted, contrary to what E9x does).

To fix this, I changed the code to first retrigger the note keeping
the volume and the envelope like it is, and then updating it if needed.

It's important to notice that, even the volume part of Rxy doesn't
apply, the value is still remembered in case a subsequent R0y is issued;
so the part handling Rxy in xm_handle_note_and_instrument isn't affected.

Another bug (more obvious) is that the multi_retrig_add lookup tables
is defined in 64th of a volume step, but the value was not scaled by 64,
so the volume change was always saturating the volume.

Co-developed with @bryc who also prepared the testcases.