Fix loading of modules with incorrect sample header size.
The XM format allows to specify the sample header size in the instrument
header. It looks like FT2-Clone and MilkyTracker both ignore this
field: they populate it correctly while saving, but they ignore it on
loading and always load the exact size of the standard sample header
(40 bytes).

We found two modules in the wild where this field was incorrect, and
libxm was not able to load them (though all trackers do).

While I did this, I realized I had not updated
xm_get_memory_needed_for_contexts for the new instrument header logic,
so I took the chance to do it.

Co-developed with @bryc.