Fix vibrato effects calculations
Previously, all the pitch effects were applied as offsets of the note.
Actually, arpeggio is defined as a note offset, while vibrato is a
period offset. The difference might be subtle but is audible in some
cases, and in general is well visible in spectrograms.

This PR changes xm_frequncy to accept both a note offset and a period
offset, and apply them correctly.

There are two attached tests.

vibrato-slow.xm uses amiga scales and plays very slow and deep vibratos
with the 4 supported waveforms (E40 sine, E41 ramp down, E42 square, E43
random -- though this last one wasn't actually supported by the original
FT2). It makes the fix clear to a listening test, though it's even more
glaring in the spectogram.

vibrato-octave.xm uses amiga scales and plays a single vibrato at
different octaves. The spectrogram makes it clear that the new formula
for vibrato in amiga frequencies is correct, though it's not easy
to hear.

Co-developed with @bryc, who also prepared the testcases.