Fix vibrato speed changes.
When vibrato is changing speed from a row to another, the previous
implementation that kept the tick count was causing a jump in the
waveform (because step = tick * speed, and if speed suddenly change,
the step jumps). Change the tick counter so that it is incremented with
the speed so that it becomes a step, and there's no more sudden change.