Add support for undefined patterns
Some modules refer to patterns that are not defined (whose index is
bigger than the number of patterns in the module). Both ft2-clone and
MilkyTracker implement this by playing an empty pattern of 64 rows
(which is the default in most editors for a new pattern), so we do
the same here.