Fix arpeggio + vibrato in linear frequencies
In #27, we refactored xm_frequency to split arpeggio (note offset) from
vibrato (period offset). Unfortunately, we did test it properly only
for amiga frequencies.

It turns out the linear frequency implementation is wrong. The note
offset cannot be added to the period, but must be added to the note
itself. In case of linear frequencies, that's trivial to do once we
get the frequency, as we can simply scale it.

This fixes binary_world.xm and so finally...

Closes #1