Cut note on period wraparound for S3M portamentos
This finally fixes PANIC.S3M.

This adds a lot of new logic in play.c, and sadly there's no easy way around it
(especially because of the different effect memory semantics).

In the future, xm_pack() or similar will preprocess and convert effects to their
simpler versions, allowing xm_analyze() to remove all of this new code anyway.